requirements.txt
hugchat==0.2.6
python-dotenv
huggingface_hub
replicate
transformers>=4.32.0
optimum>=1.12.0
auto-gptq --extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/
bardapi
accelerate
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/cu118